requirements.txt
torch==1.10.2
torchvision==0.11.3
tqdm
opencv-python
pillow
tensorboard
matplotlib
